test: move mamba_ssm-requiring puzzletron test to tests/gpu
test_resolve_descriptor_caches_dynamic_modules needs mamba_ssm (only installed
in the GPU lane) and a real Nemotron-H checkpoint, so it can only ever be
skipped in the CPU-only unit lane. Move it to tests/gpu where mamba_ssm is
available; the two mocked tests in the class stay in unit.
